requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-geoposition==0.3.0
django-model-utils==3.0.0
django-parler==1.8
django-rosetta==0.7.13
django-widget-tweaks==1.4.1
idna==2.6
microsofttranslator==0.8
polib==1.0.8
python-decouple==3.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
